+++ src/Command.ts
@@ -2,6 +2,15 @@ import { v4 as uuidv4 } from "uuid";
 import Game, { Unit, FighterType, UnitJSON } from "./Game";
 import Action from "./Action";
 import { Position, PositionJSON } from "./ObjectWithPosition";
+import MoveCommand, { MoveCommandJSON } from "./commands/MoveCommand";
+import AttackCommand, { AttackCommandJSON } from "./commands/AttackCommand";
+import SpawnCommand, { SpawnCommandJSON } from "./commands/SpawnCommand";
+import DropOffFoodCommand, {
+  DropOffFoodCommandJSON
+} from "./commands/DropOffFoodCommand";
+import PickUpFoodCommand, {
+  PickUpFoodCommandJSON
+} from "./commands/PickUpFoodCommand";
 
 export type CommandArgs = {
   position?: Position;
@@ -26,7 +35,21 @@ export type CommandClassName =
   | "DropOffFoodCommand"
   | "PickUpFoodCommand";
 
-export type CommandJSON = {
+export type CommandChildClass =
+  | MoveCommand
+  | AttackCommand
+  | SpawnCommand
+  | DropOffFoodCommand
+  | PickUpFoodCommand;
+
+export type CommandJSON =
+  | MoveCommandJSON
+  | AttackCommandJSON
+  | SpawnCommandJSON
+  | DropOffFoodCommandJSON
+  | PickUpFoodCommandJSON;
+
+export type CommandClassJSON = {
   className: CommandClassName;
   id: string;
   unit: UnitJSON;
@@ -59,9 +82,9 @@ export default class Command {
     };
   }
 
-  static fromJSON(game: Game, json: CommandJSON) {
+  static fromJSON(game: Game, json: CommandClassJSON) {
     const unit = game.lookup[json.unit.id] as Unit;
-    let args = json.args || {};
+    const args = json.args || {};
     if (args.position) {
       args.position = new Position(args.position.x, args.position.y);
     }
