+++ src/index.ts
@@ -22,9 +22,9 @@ export {
   GameGenerateProps,
   Fighter,
   FighterType,
-  Unit,
-  CommandJSON
+  Unit
 } from "./Game";
+export { CommandJSON, CommandChildClass } from "./Command";
 export { default as HQ, HQJSON } from "./HQ";
 export {
   default as ObjectWithPosition,
